LayoutTests/fast/dom/dynamic-image-with-lazy-loading-leak-expected.txt

Tests that lazy image loading doesn't cause leaks when 'img' element is created and removed dynamically.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS internals.isElementAlive(imgElementId) is true
PASS The img element didn't leak.
PASS successfullyParsed is true

TEST COMPLETE
